util done 1st Draft , waiting for improv

+++ include/util.h
@@ -1,6 +1,6 @@
 // Split and merge High byte low byte of 16 bit unsigned integer
 unsigned char *splitHLbyte(float num);
-float mergeHLbyte(unsigned char Hbyte, unsigned char Lbyte);
+unsigned int mergeHLbyte(unsigned char Hbyte, unsigned char Lbyte);
 // Encode float or uint16 into arrays of uint8
 unsigned char *Encode_bytearray(float f);
 // This is for voltage monitoring
